TF: Import all tensor values for example graphs
In order to fully resolve some tensor shapes through ops (such as concat,
transpose, shape, size, etc.), CouGr needs to be able to handle tensor values
from the graph definition. In particular, read in all the values for the TF
example graphs so they can be used in propagating shapes.