Graph: Refine tensor shape binding
In prep for some future changes, refine the way that the Graph can bind
tensor shapes. We need to bind shapes for all constants, placeholders, and
variables in the graph (source ops into the computations). Allow binding
constants, and use regular expressions to match all ops that have the same
naming conventions and op shapes.