ContextFrames: Add a concept of a context for compute blocks
Previously, we were ad-hoc handling subgraphs as a collection of ops. This
handling was sufficient for initial tests, but to add clearer subgraph
delineation for traversals, we need to make sure we include absolutely every
op that is inside a context (e.g., context = while loop or if statement).
This PR adds the context to track the EnterOps that control/gate tensors
flowing into the context and makes it easier to detect all the downstream ops
that are within the context.