ContextFrames: Add a concept of a context for compute blocks (#25)
* ContextFrames: Add a concept of a context for compute blocks

Previously, we were ad-hoc handling subgraphs as a collection of ops. This
handling was sufficient for initial tests, but to add clearer subgraph
delineation for traversals, we need to make sure we include absolutely every
op that is inside a context (e.g., context = while loop or if statement).
This PR adds the context to track the EnterOps that control/gate tensors
flowing into the context and makes it easier to detect all the downstream ops
that are within the context.

* Associate ContextFrame with SubgraphOp
* Update full tests

The while-loop subgraphs of some tests added a few new ops since we are
now correctly traversing subgraphs from all associated EnterOps to all
the associated ExitOps and MergeOps. As a result, some ops are now added
into the loops and so counted once per iteration rather than just a
single time. Small adjustments to tests.